# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KPI tizimi API - Kattaqo'rg'on tumani suv yetkazib berish xizmati
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  manager: "manager",
  viewer: "viewer",
} as const;

export interface User {
  id: number;
  username: string;
  fullName: string;
  role: UserRole;
  departmentId?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Department {
  id: number;
  name: string;
  code: string;
  description?: string | null;
  headName?: string | null;
  employeeCount: number;
  createdAt: string;
}

export interface CreateDepartmentBody {
  name: string;
  code: string;
  description?: string | null;
  headName?: string | null;
}

export type EmployeeStatus =
  (typeof EmployeeStatus)[keyof typeof EmployeeStatus];

export const EmployeeStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Employee {
  id: number;
  fullName: string;
  position: string;
  departmentId: number;
  departmentName?: string | null;
  phone?: string | null;
  email?: string | null;
  hireDate?: string | null;
  status: EmployeeStatus;
  averageScore?: number | null;
  createdAt: string;
}

export type CreateEmployeeBodyStatus =
  (typeof CreateEmployeeBodyStatus)[keyof typeof CreateEmployeeBodyStatus];

export const CreateEmployeeBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateEmployeeBody {
  fullName: string;
  position: string;
  departmentId: number;
  phone?: string | null;
  email?: string | null;
  hireDate?: string | null;
  status: CreateEmployeeBodyStatus;
}

export interface KpiCategory {
  id: number;
  name: string;
  description?: string | null;
  weight: number;
  color?: string | null;
  indicatorCount: number;
  createdAt: string;
}

export interface CreateKpiCategoryBody {
  name: string;
  description?: string | null;
  weight: number;
  color?: string | null;
}

export interface KpiIndicator {
  id: number;
  name: string;
  description?: string | null;
  categoryId: number;
  categoryName?: string | null;
  unit?: string | null;
  targetValue?: number | null;
  weight: number;
  createdAt: string;
}

export interface CreateKpiIndicatorBody {
  name: string;
  description?: string | null;
  categoryId: number;
  unit?: string | null;
  targetValue?: number | null;
  weight: number;
}

export interface Evaluation {
  id: number;
  employeeId: number;
  employeeName?: string | null;
  departmentName?: string | null;
  indicatorId: number;
  indicatorName?: string | null;
  categoryName?: string | null;
  period: string;
  score: number;
  maxScore: number;
  comment?: string | null;
  evaluatorId?: number | null;
  evaluatorName?: string | null;
  createdAt: string;
}

export interface CreateEvaluationBody {
  employeeId: number;
  indicatorId: number;
  period: string;
  score: number;
  maxScore: number;
  comment?: string | null;
}

export interface DashboardSummary {
  totalEmployees: number;
  totalDepartments: number;
  totalKpiIndicators: number;
  averageScore: number;
  evaluationsThisMonth: number;
  highPerformers: number;
  lowPerformers: number;
  currentPeriod: string;
}

export interface EmployeeRanking {
  employeeId: number;
  fullName: string;
  position: string;
  departmentName?: string | null;
  averageScore: number;
  scorePercentage: number;
  rank: number;
}

export interface DepartmentScore {
  departmentId: number;
  departmentName: string;
  averageScore: number;
  scorePercentage: number;
  employeeCount: number;
}

export interface MonthlyTrend {
  month: string;
  averageScore: number;
  evaluationCount: number;
}

export type WorkPlanTaskStatus =
  (typeof WorkPlanTaskStatus)[keyof typeof WorkPlanTaskStatus];

export const WorkPlanTaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface WorkPlanTask {
  id: number;
  planId: number;
  title: string;
  description?: string | null;
  deadline?: string | null;
  completionPercentage: number;
  status: WorkPlanTaskStatus;
  createdAt: string;
}

export type CreateWorkPlanTaskBodyStatus =
  (typeof CreateWorkPlanTaskBodyStatus)[keyof typeof CreateWorkPlanTaskBodyStatus];

export const CreateWorkPlanTaskBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface CreateWorkPlanTaskBody {
  title: string;
  description?: string | null;
  deadline?: string | null;
  completionPercentage: number;
  status: CreateWorkPlanTaskBodyStatus;
}

export type WorkPlanStatus =
  (typeof WorkPlanStatus)[keyof typeof WorkPlanStatus];

export const WorkPlanStatus = {
  draft: "draft",
  submitted: "submitted",
  approved: "approved",
  completed: "completed",
} as const;

export interface WorkPlan {
  id: number;
  employeeId: number;
  employeeName?: string | null;
  departmentName?: string | null;
  period: string;
  title: string;
  description?: string | null;
  status: WorkPlanStatus;
  approvedById?: number | null;
  approvedByName?: string | null;
  approveComment?: string | null;
  tasks: WorkPlanTask[];
  taskCount: number;
  completedTaskCount: number;
  overallProgress: number;
  createdAt: string;
}

export interface CreateWorkPlanBody {
  employeeId: number;
  period: string;
  title: string;
  description?: string | null;
}

export type ListEmployeesParams = {
  departmentId?: number;
  search?: string;
};

export type ListKpiIndicatorsParams = {
  categoryId?: number;
};

export type ListEvaluationsParams = {
  employeeId?: number;
  period?: string;
  departmentId?: number;
};

export type GetTopEmployeesParams = {
  period?: string;
  limit?: number;
};

export type GetDepartmentScoresParams = {
  period?: string;
};

export type GetRecentEvaluationsParams = {
  limit?: number;
};

export type ListWorkPlansParams = {
  employeeId?: number;
  period?: string;
  departmentId?: number;
  status?: string;
};

export type ApproveWorkPlanBody = {
  comment?: string | null;
};